Added meter edit view

+++ app/src/main/java/com/tentlers/mngapp/data/tables/meters/MetersListObj.java
@@ -10,8 +10,7 @@ import androidx.room.Ignore;
 
 public class MetersListObj {
     @Ignore
-    public int houseId;
-
+    public long houseId;
     public long meterId;
     public String houseName;
     public String roomName;
